+++ crates/formatter/src/formatter.rs
@@ -431,7 +431,6 @@ impl VerylWalker for Formatter {
         match &*arg.type_group {
             TypeGroup::BuiltinType(x) => self.builtin_type(&x.builtin_type),
             TypeGroup::ScopedIdentifier(x) => self.scoped_identifier(&x.scoped_identifier),
-            TypeGroup::ModportIdentifier(x) => self.modport_identifier(&x.modport_identifier),
         };
         self.space(1);
         for x in &arg.type_list {