+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased](https://github.com/dalance/veryl/compare/v0.1.14...Unreleased) - ReleaseDate
 
+* [Changed] modport separator from `.` to `::`
+
 ## [v0.1.14](https://github.com/dalance/veryl/compare/v0.1.13...v0.1.14) - 2023-01-12
 
 ## [v0.1.13](https://github.com/dalance/veryl/compare/v0.1.12...v0.1.13) - 2023-01-10